requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
transformers==4.35.0
torch==2.1.0
python-multipart==0.0.6
websockets==11.0.3
langchain==0.0.330
openai==1.3.0
tiktoken==0.5.1 